make a change to the investor signature payload.
Prior to this diff an investor committed to the other investor's address. This was bad because it means the investor needs to know
the address of their counterparty when they make an Order. In this diff we remove the counterparty investor from the signature, which
frees up the offchain logic so now the investor signatures can be generated on Order creation/update, rather than after Trade matching.
This reduces the number of round trips needed in the DvP flow.